requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.1.0
boto3==1.28.16
botocore==1.31.16
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==41.0.2
daphne==4.0.0
dataclasses-json==0.5.13
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-channels==0.7.0
django-crispy-forms==2.0
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
frozenlist==1.4.0
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
joblib==1.3.1
kombu==5.3.1
langchain==0.0.247
langsmith==0.0.15
Markdown==3.4.3
marshmallow==3.20.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pgvector==0.2.0
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
PyJWT==2.7.0
PyMuPDF==1.22.5
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.11.1
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
tenacity==8.2.2
threadpoolctl==3.2.0
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
yarl==1.9.2
zope.interface==6.0
